+++ src/Microsoft.Diagnostics.Tools.RuntimeClient/Eventing/EventPipeClient.cs
@@ -67,43 +67,6 @@ namespace Microsoft.Diagnostics.Tools.RuntimeClient
         }
 
         /// <summary>
-        /// Get all active process ports
-        /// </summary>
-        /// <returns>
-        /// A collection of string identifiers that show the current processes
-        /// </returns>
-        public static string GetProcessStatus()
-        {
-            try
-            {
-                StringBuilder sb = new StringBuilder();
-                var processes = ListAvailablePorts()
-                    .Select(GetProcessById)
-                    .Where(process => process != null)
-                    .OrderBy(process => process.ProcessName)
-                    .ThenBy(process => process.Id);
-
-
-                foreach (var process in processes)
-                {
-                    try
-                    {
-                        sb.Append($"{process.Id, 10} {process.ProcessName, -10} {process.MainModule.FileName}\n");
-                    }
-                    catch (Exception)
-                    {
-                        sb.Append($"{process.Id, 10} {process.ProcessName, -10} [Elevated process - cannot determine path]\n");
-                    }
-                }
-                return sb.ToString();
-            }
-            catch (Exception ex)
-            {
-                return ex.ToString();
-            }
-        }
-
-        /// <summary>
         /// Start trace collection.
         /// </summary>
         /// <param name="processId">Runtime process to trace</param>
@@ -178,18 +141,6 @@ namespace Microsoft.Diagnostics.Tools.RuntimeClient
                     return 0;
                 default:
                     return 0;
-            }
-        }
-
-        private static Process GetProcessById(int processId)
-        {
-            try
-            {
-                return Process.GetProcessById(processId);
-            }
-            catch (ArgumentException)
-            {
-                return null;
             }
         }
