+++ src/Tools/dotnet-counters/dotnet-counters.csproj
@@ -12,6 +12,7 @@
 
   <ItemGroup>
     <Compile Include="$(MSBuildThisFileDirectory)..\dotnet-trace\Extensions.cs" Link="Extensions.cs" />
+    <Compile Include="..\Common\Commands\ListProcesses.cs" Link="ListProcesses.cs" />
   </ItemGroup>
 
   <ItemGroup>
