+++ src/Tools/dotnet-dump/dotnet-dump.csproj
@@ -16,6 +16,10 @@
     <PackageReference Include="Microsoft.Diagnostics.Runtime" Version="$(MicrosoftDiagnosticsRuntimeVersion)" />
     <PackageReference Include="Microsoft.SymbolStore" Version="$(MicrosoftSymbolStoreVersion)" />
   </ItemGroup>
+
+  <ItemGroup>
+    <Compile Include="..\Common\Commands\ListProcesses.cs" Link="ListProcesses.cs" />
+  </ItemGroup>
   
   <ItemGroup>
     <ProjectReference Include="$(MSBuildThisFileDirectory)..\..\Microsoft.Diagnostics.Repl\Microsoft.Diagnostics.Repl.csproj" />
